Implement Windows support
Originally my plan was to wait until some random user contributes
support for Windows, because I don't really care about that platform.
But let's face it: The odds of finding a skilled developer that is still
on Windows are already pretty low, but the odds of finding a Windows
developer that is good in FOSS and that uses vim on Windows of all
platforms are pretty much non-existent.

Together with the outlook on the hacky task of calling a bunch of Win32
API through Powershell - a single dumpster fire of a shell - it became
clear that there was no other way than to take matters in my own hands
and implement this myself.

And oh boy, was I not prepared for this epitome of cursedness.
Even the MacOS support that literally has to compile Swift programs on
the fly looks benign in comparison to this feast of Win32 API.

And what is up with Windows' fetish for handles everywhere? Oh, you want
to get a handle on a registry key? You better pass another useless root
handle for HKEY_CURRENT_USER so that we can return you yet another
handle for the key that you are really interested in.

OK before this ends in a page-long rant about what the hell is all
wrong with development on Windows, I'd rather quickly explain this
patch:

There luckily is a registry key that returns the user's dark mode
preference:

HKEY_CURRENT_USER\Software\Microsoft\Windows\CurrentVersion\Themes\Personalize\AppsUseLightTheme

Querying this key is easy with "reg query". Subscribing to changes
without polling is a bit more involved though. There doesn't really
exist a pure Powershell way to do this, so we abuse Powershell's ability
to call into Win32 API via signatures as strings (lol) to accomplish
this with interrupts (i.e. we get notified as soon something changes,
instead of having to poll).
For that we use another event struct and WaitForSingleObject that blocks
our thread until the next event is ready.

The windows platform itself can be detected by checking for the win32
feature, which is also set on 64bit Windows.
We must also expand <script> instead of <sfile> (which is used in the
MacOS platform), because it is expanded inside the function call instead
of free without function in the script file.

Another thing to watch out for is the whole CRLF bullshit, e.g. most vim
methods that convert from lines to arrays really only trim the line
feed, not the carriage return.
There also is some weird execution policy bullshit that prevents us from
running the script directly, but we can easily circumvent it just by
passing a parameter, wtf.
There is just so much bullshit when developing on Windows that gets into
your way in little but very annoying ways. Of course I also got a
forced update from this meme operating system as soon as I shut down...

Alright, this is enough Windows for me, can't wait to start you up again
in another 2 years.

Fixes #3