Avoid calling callbacks twice
Some platforms such as the newly added Windows support call the same
setting multiple times in a row (don't ask me why the registry change
even is triggered twice on Windows).
Instead of invoking the autocmd multiple times, let's quit early if we
already set the background to the correct value in the first callback.

The additional check for "&background == s:background" is needed,
because otherwise we permanently disable the plugin forever, if
&background ever diverges from our "cached" background. This might be
wanted sometimes (e.g. if the user manually sets it and then later on
the system dark mode changes coincidentally), but I fear that some
plugins might be deliberately setting background to a fixed value
without thinking too much about the consequences, and if that happens
after this plugin was loaded, it would permanently disable this plugin.
Thus it is safer to keep this check.

On the other hand we always want to trigger the autocmd on the first
time (even if background already is the same value), because the user
might want to set some other related settings (e.g. fzf colors) to some
value depending on the callback. This is why the check for the cached
variable is required.