Escape all system command filenames
In the macos implementation it is possible that spaces appear in the
absolute file path, so we have to escape it for all system command uses.

All other platform implementations are not able to end up with spaces
in system command arguments (most of them are hardcoded and do not
depend on the directory where vim is installed).

Fixes #13